Remove record_against_rival from player show page
I thought that displaying all matches was what took so long to load the
player show page, but I think it actually has something to do with the
way the rival stats are calculated. I'm probably performing a bunch of
unnecessary calculations with those functions. For now, this should
speed up the player show page considerably, and the record between
opponents should still be available on the match show page.